[nrf noup] net: sockets: Move NCS additions to a separate header
Move NCS specific socket API additions to a separate header, for easier
management of these in the future and less conflicts during upmerges.

To be squashed into 745826d7816a4672d09725d691ff6a6a8bdad3a4.
